+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/execution/sam/SamInvokeRunner.kt
@@ -3,7 +3,6 @@
 
 package software.aws.toolkits.jetbrains.services.lambda.execution.sam
 
-import com.intellij.execution.ExecutionException
 import com.intellij.execution.configurations.RunProfile
 import com.intellij.execution.configurations.RunProfileState
 import com.intellij.execution.configurations.RunnerSettings
@@ -11,8 +10,8 @@ import com.intellij.execution.executors.DefaultDebugExecutor
 import com.intellij.execution.executors.DefaultRunExecutor
 import com.intellij.execution.runners.AsyncProgramRunner
 import com.intellij.execution.runners.ExecutionEnvironment
+import com.intellij.execution.runners.RunContentBuilder
 import com.intellij.execution.ui.RunContentDescriptor
-import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.project.Project
@@ -26,7 +25,6 @@ import software.aws.toolkits.core.lambda.validOrNull
 import software.aws.toolkits.core.telemetry.DefaultMetricEvent.Companion.METADATA_INVALID
 import software.aws.toolkits.core.utils.getLogger
 import software.aws.toolkits.core.utils.tryOrNull
-import software.aws.toolkits.core.utils.warn
 import software.aws.toolkits.jetbrains.core.AwsResourceCache
 import software.aws.toolkits.jetbrains.services.PathMapping
 import software.aws.toolkits.jetbrains.services.lambda.Lambda
@@ -92,158 +90,15 @@ class SamInvokeRunner : AsyncProgramRunner<RunnerSettings>() {
     }
 
     override fun execute(environment: ExecutionEnvironment, state: RunProfileState): Promise<RunContentDescriptor?> {
-        FileDocumentManager.getInstance().saveAllDocuments()
-
-        val buildingPromise = AsyncPromise<RunContentDescriptor?>()
-        val samState = state as SamRunningState
-        val lambdaSettings = samState.settings
-        val lambdaBuilder = LambdaBuilder.getInstance(lambdaSettings.runtimeGroup)
-
-        val buildLambdaRequest = when (lambdaSettings) {
-            is TemplateRunSettings ->
-                buildLambdaFromTemplate(
-                    lambdaSettings.templateFile,
-                    lambdaSettings.logicalId,
-                    lambdaSettings.samOptions
-                )
-            is ImageTemplateRunSettings ->
-                buildLambdaFromTemplate(
-                    lambdaSettings.templateFile,
-                    lambdaSettings.logicalId,
-                    lambdaSettings.samOptions
-                )
-            is HandlerRunSettings ->
-                buildLambdaFromHandler(
-                    lambdaBuilder,
-                    environment.project,
-                    lambdaSettings
-                )
+        val runPromise = AsyncPromise<RunContentDescriptor?>()
+        val runContentDescriptor = state.execute(environment.executor, this)?.let {
+            RunContentBuilder(it, environment).showRunContent(environment.contentToReuse)
         }
 
-        val buildWorkflow = buildWorkflow(environment, buildLambdaRequest, lambdaSettings.samOptions)
-        buildWorkflow.onSuccess = { context ->
-            val builtLambda = context.getRequiredAttribute(BuildLambda.BUILT_LAMBDA)
-            samState.builtLambda = builtLambda
-            samState.pathMappings = createPathMappings(lambdaBuilder, lambdaSettings, buildLambdaRequest)
+        runPromise.setResult(runContentDescriptor)
 
-            samState.runner.run(environment, samState)
-                .onSuccess {
-                    buildingPromise.setResult(it)
-                    reportMetric(lambdaSettings, Result.Succeeded, environment.isDebug())
-                }.onError {
-                    buildingPromise.setError(it)
-                    reportMetric(lambdaSettings, Result.Failed, environment.isDebug())
-                }
-        }
-        buildWorkflow.onError = {
-            // Remap to ExecutionException so our run configuration fails properly
-            // instead of showing up as an IDE Fatal error
-            val exception = if (it is ExecutionException) {
-                it
-            } else {
-                ExecutionException(it)
-            }
-            LOG.warn(it) { "Failed to create Lambda package" }
-            buildingPromise.setError(exception)
-            reportMetric(lambdaSettings, Result.Failed, environment.isDebug())
-        }
-
-        buildWorkflow.startExecution()
-
-        return buildingPromise
+        return runPromise
     }
-
-    private fun createPathMappings(lambdaBuilder: LambdaBuilder, settings: LocalLambdaRunSettings, buildRequest: BuildRequest): List<PathMapping> {
-        val defaultPathMappings = lambdaBuilder.defaultPathMappings(buildRequest.template, buildRequest.logicalId, buildRequest.buildDir)
-        return if (settings is ImageTemplateRunSettings) {
-            // This needs to be a bit smart. If a user set local path matches a default path, we need to make sure that is the one set
-            // by removing the default set one.
-            val userMappings = settings.pathMappings.map { PathMapping(it.localRoot, it.remoteRoot) }
-            userMappings + defaultPathMappings.filterNot { defaultMapping -> userMappings.any { defaultMapping.localRoot == it.localRoot } }
-        } else {
-            defaultPathMappings
-        }
-    }
-
-    private fun reportMetric(lambdaSettings: LocalLambdaRunSettings, result: Result, isDebug: Boolean) {
-        val account = AwsResourceCache.getInstance()
-            .getResourceIfPresent(StsResources.ACCOUNT, lambdaSettings.connection)
-
-        LambdaTelemetry.invokeLocal(
-            metadata = MetricEventMetadata(
-                awsAccount = account ?: METADATA_INVALID,
-                awsRegion = lambdaSettings.connection.region.id
-            ),
-            debug = isDebug,
-            runtime = TelemetryRuntime.from(
-                when (lambdaSettings) {
-                    is ZipSettings -> {
-                        lambdaSettings.runtime.toString()
-                    }
-                    is ImageSettings -> {
-                        lambdaSettings.imageDebugger.id
-                    }
-                    else -> {
-                        ""
-                    }
-                }
-            ),
-            lambdaPackageType = if (lambdaSettings is ImageTemplateRunSettings) LambdaPackageType.Image else LambdaPackageType.Zip,
-            result = result
-        )
-    }
-
-    // TODO: We actually probably want to split this for image templates and handler templates to enable build env vars for handler based
-    private fun buildLambdaFromTemplate(templateFile: VirtualFile, logicalId: String, samOptions: SamOptions): BuildRequest {
-        val templatePath = Paths.get(templateFile.path)
-        val buildDir = templatePath.resolveSibling(".aws-sam").resolve("build")
-
-        return BuildRequest(templatePath, logicalId, emptyMap(), buildDir)
-    }
-
-    private fun buildLambdaFromHandler(lambdaBuilder: LambdaBuilder, project: Project, settings: HandlerRunSettings): BuildRequest {
-        val samOptions = settings.samOptions
-        val runtime = settings.runtime
-        val handler = settings.handler
-
-        val element = Lambda.findPsiElementsForHandler(project, runtime, handler).first()
-        val module = getModule(element.containingFile)
-
-        val buildDirectory = lambdaBuilder.getBuildDirectory(module)
-        val dummyTemplate = buildDirectory.parent.resolve("temp-template.yaml")
-        val dummyLogicalId = "Function"
-
-        SamTemplateUtils.writeDummySamTemplate(
-            tempFile = dummyTemplate,
-            logicalId = dummyLogicalId,
-            runtime = runtime,
-            handler = handler,
-            timeout = settings.timeout,
-            memorySize = settings.memorySize,
-            codeUri = lambdaBuilder.handlerBaseDirectory(module, element).toAbsolutePath().toString(),
-            envVars = settings.environmentVariables
-        )
-
-        return BuildRequest(
-            dummyTemplate,
-            dummyLogicalId,
-            lambdaBuilder.additionalBuildEnvironmentVariables(module, samOptions),
-            buildDirectory
-        )
-    }
-
-    private fun buildWorkflow(environment: ExecutionEnvironment, buildRequest: BuildRequest, samOptions: SamOptions): StepExecutor {
-        val buildStep = BuildLambda(buildRequest.template, buildRequest.logicalId, buildRequest.buildDir, buildRequest.buildEnvVars, samOptions)
-
-        return StepExecutor(environment.project, message("sam.build.running"), StepWorkflow(ValidateDocker(), buildStep), environment.executionId.toString())
-    }
-
-    private fun getModule(psiFile: PsiFile): Module = ModuleUtil.findModuleForFile(psiFile)
-        ?: throw IllegalStateException("Failed to locate module for $psiFile")
-
-    private fun ExecutionEnvironment.isDebug(): Boolean = (executor.id == DefaultDebugExecutor.EXECUTOR_ID)
-
-    private data class BuildRequest(val template: Path, val logicalId: String, val buildEnvVars: Map<String, String>, val buildDir: Path)
 
     private companion object {
         val LOG = getLogger<SamInvokeRunner>()
