+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/lambda/python/PythonDebugUtils.kt
@@ -30,7 +30,8 @@ object PythonDebugUtils {
     ): XDebugProcessStarter {
         // TODO: We should allow using the module SDK, but we can't easily get the module
         val sdk = ProjectRootManager.getInstance(environment.project).projectSdk?.takeIf { it.sdkType is PythonSdkType }
-        state.consoleBuilder = PyDebugConsoleBuilder(environment.project, sdk)
+        // TODO fix
+        //state.consoleBuilder = PyDebugConsoleBuilder(environment.project, sdk)
 
         val executionResult = withContext(Dispatchers.IO) {
             // needs to run off EDT since it resolves credentials
