+++ resources/resources/software/aws/toolkits/resources/MessagesBundle.properties
@@ -476,6 +476,7 @@ lambda.debug.attach.error=Debugger attach error
 lambda.debug.attach.fail=Unable to attach to debugger process.
 lambda.debug.docker.not_connected=Running/Debugging Lambdas locally requires Docker. Please make sure it is installed and running
 lambda.debug.process.start.timeout=Timeout waiting for Debugger Worker process to start
+lambda.debug.step.start_sam=Running SAM local invoke
 lambda.debug.waiting=Waiting for debugger...
 lambda.execute.function_error=Function error: {0}
 lambda.execute.invoke=Invoking Lambda function: {0}
